Rename InteractiveAuthFlowStepUriMapper to InteractiveFlowStepUriMapper (#276)
Drop the Auth qualifier and move the mapper out of the api.controller.flow.auth package up to api.controller.flow: the step vocabulary (InteractiveFlowStep) and its URI mapping are generic to the interactive flow, not specific to the OAuth2 authorization consumer.
